Restructured project.
* Flattened the structure.
* Congregated functionality.
